chore: move engine out of celery tasks
